+++ app/controllers/api/payments_controller.rb
@@ -1,15 +1,17 @@
 module Api
   class PaymentsController < ApplicationController
+    before_action :check_request_payment, only: %i[show update destroy]
+
     def create
-      attributes = params.permit(:date, :content, :price, :payment_type, categories: [])
+      required_param_keys = %i[payment_type date content categories price]
+      check_absent_param(create_params, required_param_keys)
 
-      absent_keys = create_params - attributes.keys.map(&:to_sym)
-      error_codes = absent_keys.map {|key| "absent_param_#{key}" }
-      raise BadRequest, error_codes unless absent_keys.empty?
-
-      @payment = Payment.new(attributes.except(:categories))
-      @payment.categories << attributes[:categories].map do |category_name|
+      @payment = Payment.new(create_params.except(:categories, :tags))
+      @payment.categories = create_params[:categories].map do |category_name|
         Category.find_or_initialize_by(name: category_name)
+      end
+      @payment.tags = Array.wrap(create_params[:tags]).map do |tag_name|
+        Tag.find_or_initialize_by(name: tag_name)
       end
 
       if @payment.save
@@ -21,14 +23,11 @@ module Api
     end
 
     def show
-      @payment = Payment.find_by(params.permit(:id))
-      raise NotFound unless @payment
-
       render status: :ok, template: 'payments/payment'
     end
 
     def index
-      query = Query.new(params.permit(*index_params))
+      query = PaymentQuery.new(params.permit(*index_params))
       if query.valid?
         query_params = index_params - %i[page per_page sort order]
         @payments = query_params.inject(Payment.all) do |payments, key|
@@ -43,35 +42,45 @@ module Api
     end
 
     def update
-      @payment = Payment.find_by(params.permit(:id))
-      raise NotFound unless @payment
-
       attributes = params.permit(:date, :content, :price, :payment_type, categories: [])
       if attributes[:categories]
-        @payment.categories = attributes[:categories].map do |category_name|
+        payment.categories = attributes[:categories].map do |category_name|
           Category.find_or_create_by(name: category_name)
         end
       end
 
-      if @payment.update(attributes.except(:categories))
+      if payment.update(attributes.except(:categories))
         render status: :ok, template: 'payments/payment'
       else
-        error_codes = @payment.errors.messages.keys.map {|key| "invalid_param_#{key}" }
+        error_codes = payment.errors.messages.keys.map {|key| "invalid_param_#{key}" }
         raise BadRequest, error_codes
       end
     end
 
     def destroy
-      @payment = Payment.find_by(params.permit(:id)).try(:destroy)
-      raise NotFound unless @payment
-
+      payment.destroy
       head :no_content
     end
 
     private
 
+    def check_request_payment
+      raise NotFound unless payment
+    end
+
+    def payment
+      @payment ||= Payment.find_by(request.path_parameters.slice(:payment_id))
+    end
+
     def create_params
-      %i[payment_type date content categories price]
+      @create_params ||= request.request_parameters.slice(
+        :payment_type,
+        :date,
+        :content,
+        :categories,
+        :tags,
+        :price,
+      )
     end
 
     def index_params