+++ app/controllers/payments_controller.rb
@@ -1,18 +1,14 @@
 class PaymentsController < ApplicationController
   def index
-    @search_form = Query.new(params.permit(*index_params))
-    per_page = request.query_parameters[:per_page]
-    raise BadRequest, 'invalid_param_per_page' if per_page and per_page !~ /\A\d*\z/
+    @search_form = PaymentQuery.new(index_param)
 
-    @per_page = per_page ? per_page.to_i : Kaminari.config.default_per_page
     if @search_form.valid?
-      @payments = index_params.inject(Payment.all) do |payments, key|
+      @payments = scope_param.keys.inject(Payment.all) do |payments, key|
         value = @search_form.send(key)
         value ? payments.send(key, value) : payments
       end
       @payment = Payment.new
-      @payments = @payments.order(date: :desc).page(params[:page]).per(@per_page)
-      @dictionary = Dictionary.new
+      @payments = @payments.order(date: :desc).page(params[:page]).per(per_page)
       render status: :ok
     else
       error_codes = @search_form.errors.messages.keys.map {|key| "invalid_param_#{key}" }
@@ -22,10 +18,39 @@ class PaymentsController < ApplicationController
 
   private
 
-  def index_params
-    %i[
-      payment_type date_before date_after content_equal content_include category
-      price_upper price_lower
-    ]
+  def index_param
+    @index_param ||= request.query_parameters.slice(
+      :payment_type,
+      :date_before,
+      :date_after,
+      :content_equal,
+      :content_include,
+      :category,
+      :tag,
+      :price_upper,
+      :price_lower,
+      :page,
+      :per_page,
+      :sort,
+      :order,
+    )
+  end
+
+  def scope_param
+    @scope_param ||= index_param.except(
+      :page,
+      :per_page,
+      :sort,
+      :order,
+    )
+  end
+
+  def per_page
+    return @per_page if @per_page
+
+    per_page = index_param[:per_page] || Kaminari.config.default_per_page
+    raise BadRequest, 'invalid_param_per_page' unless per_page.to_s.match?('\A\d*\z')
+
+    @per_page = per_page.to_i
   end
 end