+++ app/models/category.rb
@@ -4,5 +4,15 @@ class Category < ApplicationRecord
   has_many :category_dictionaries, dependent: :destroy
   has_many :dictionaries, through: :category_dictionaries
 
-  validates :name, presence: {message: 'absent'}
+  validates :category_id, :name,
+            presence: {message: 'absent'}
+  validates :category_id,
+            format: {with: /\A[0-9a-f]{32}\z/, message: 'invalid'},
+            allow_nil: true
+
+  scope :name_include, ->(name) { where('name REGEXP ?', ".*#{name}.*") }
+
+  after_initialize if: :new_record? do |category|
+    category.category_id = SecureRandom.hex
+  end
 end