+++ app/models/payment.rb
@@ -8,10 +8,15 @@ class Payment < ApplicationRecord
 
   has_many :category_payments, dependent: :destroy
   has_many :categories, through: :category_payments
+  has_many :payment_tags, dependent: :destroy
+  has_many :tags, through: :payment_tags
 
-  validates :payment_type, :content, :price,
+  validates :payment_id, :payment_type, :content, :price,
             presence: {message: 'absent'}
   validates :date, presence: {message: 'invalid'}
+  validates :payment_id,
+            format: {with: /\A[0-9a-f]{32}\z/, message: 'invalid'},
+            allow_nil: true
   validates :payment_type,
             inclusion: {in: PAYMENT_TYPES, message: 'invalid'},
             allow_nil: true
@@ -31,6 +36,13 @@ class Payment < ApplicationRecord
   scope :category, lambda {|category|
     joins(:categories).where('categories.name' => category.split(','))
   }
+  scope :tag, lambda {|tag|
+    joins(:tags).where('tags.name' => tag.split(','))
+  }
   scope :price_upper, ->(price) { where('price >= ?', price) }
   scope :price_lower, ->(price) { where('price <= ?', price) }
+
+  after_initialize if: :new_record? do |payment|
+    payment.payment_id = SecureRandom.hex
+  end
 end