+++ db/schema.rb
@@ -10,13 +10,15 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20190701104957) do
+ActiveRecord::Schema.define(version: 20200208070525) do
 
   create_table "categories", force: :cascade, options: "ENGINE=InnoDB DEFAULT CHARSET=utf8" do |t|
+    t.string   "category_id"
     t.string   "name",                      null: false
     t.text     "description", limit: 65535
     t.datetime "created_at",                null: false
     t.datetime "updated_at",                null: false
+    t.index ["category_id"], name: "index_categories_on_category_id", unique: true, using: :btree
     t.index ["name"], name: "index_categories_on_name", unique: true, using: :btree
   end
 
@@ -41,20 +43,43 @@ ActiveRecord::Schema.define(version: 20190701104957) do
   end
 
   create_table "dictionaries", force: :cascade, options: "ENGINE=InnoDB DEFAULT CHARSET=utf8" do |t|
-    t.string   "phrase",     null: false
-    t.string   "condition",  null: false
-    t.datetime "created_at", null: false
-    t.datetime "updated_at", null: false
+    t.string   "dictionary_id"
+    t.string   "phrase",        null: false
+    t.string   "condition",     null: false
+    t.datetime "created_at",    null: false
+    t.datetime "updated_at",    null: false
+    t.index ["dictionary_id"], name: "index_dictionaries_on_dictionary_id", unique: true, using: :btree
     t.index ["phrase", "condition"], name: "index_dictionaries_on_phrase_and_condition", unique: true, using: :btree
   end
 
+  create_table "payment_tags", force: :cascade, options: "ENGINE=InnoDB DEFAULT CHARSET=utf8" do |t|
+    t.integer  "payment_id", null: false
+    t.integer  "tag_id",     null: false
+    t.datetime "created_at", null: false
+    t.datetime "updated_at", null: false
+    t.index ["payment_id", "tag_id"], name: "index_payment_tags_on_payment_id_and_tag_id", unique: true, using: :btree
+    t.index ["payment_id"], name: "index_payment_tags_on_payment_id", using: :btree
+    t.index ["tag_id"], name: "index_payment_tags_on_tag_id", using: :btree
+  end
+
   create_table "payments", force: :cascade, options: "ENGINE=InnoDB DEFAULT CHARSET=utf8" do |t|
+    t.string   "payment_id"
     t.string   "payment_type"
     t.date     "date"
     t.string   "content"
     t.integer  "price"
     t.datetime "created_at",   null: false
     t.datetime "updated_at",   null: false
+    t.index ["payment_id"], name: "index_payments_on_payment_id", unique: true, using: :btree
+  end
+
+  create_table "tags", force: :cascade, options: "ENGINE=InnoDB DEFAULT CHARSET=utf8" do |t|
+    t.string   "tag_id"
+    t.string   "name",       default: "", null: false
+    t.datetime "created_at",              null: false
+    t.datetime "updated_at",              null: false
+    t.index ["name"], name: "index_tags_on_name", unique: true, using: :btree
+    t.index ["tag_id"], name: "index_tags_on_tag_id", unique: true, using: :btree
   end
 
 end