+++ spec/controllers/api/payment/show_spec.rb
@@ -3,9 +3,10 @@
 require 'rails_helper'
 
 describe Api::PaymentsController, type: :controller do
-  shared_context '収支情報を取得する' do |id|
+  shared_context '収支情報を取得する' do |payment_id|
     before(:all) do
-      res = client.get("/api/payments/#{id}")
+      payment_id ||= @payment_id
+      res = client.get("/api/payments/#{payment_id}")
       @response_status = res.status
       @response_body = JSON.parse(res.body) rescue res.body
     end
@@ -17,17 +18,26 @@ describe Api::PaymentsController, type: :controller do
     payment = PaymentHelper.test_payment[:income]
     before(:all) do
       categories = payment[:categories].map do |category_name|
-        Category.find_by(name: category_name).slice(:id, :name, :description)
+        Category.find_by(name: category_name).slice(:category_id, :name, :description)
       end
-      @body = payment.merge(categories: categories).deep_stringify_keys
+      tags = payment[:tags].map do |tag_name|
+        Tag.find_by(name: tag_name).slice(:tag_id, :name)
+      end
+
+      @payment_id = Payment.find(payment[:id]).payment_id
+      @body = payment.except(:id).merge(
+        payment_id: @payment_id,
+        categories: categories,
+        tags: tags,
+      ).deep_stringify_keys
     end
-    include_context '収支情報を取得する', payment[:id]
+    include_context '収支情報を取得する'
     it_behaves_like 'レスポンスが正しいこと'
   end
 
   describe '異常系' do
     context '存在しないidを指定した場合' do
-      include_context '収支情報を取得する', 100
+      include_context '収支情報を取得する', 'not_exist'
       it_behaves_like 'レスポンスが正しいこと', status: 404, body: ''
     end
   end