+++ spec/factories/payment.rb
@@ -1,9 +1,11 @@
 FactoryBot.define do
   factory :payment do
+    payment_id { SecureRandom.hex }
     payment_type { 'income' }
     date { '1000-01-01' }
     content { 'test' }
     categories { [FactoryBot.create(:category)] }
+    tags { [FactoryBot.create(:tag)] }
     price { 1000 }
   end
 end