+++ spec/models/category_spec.rb
@@ -6,6 +6,7 @@ describe Category, type: :model do
   describe '#validates' do
     describe '正常系' do
       valid_attribute = {
+        category_id: ['0' * 32],
         name: %w[test],
         description: ['test', nil],
       }
@@ -14,7 +15,12 @@ describe Category, type: :model do
     end
 
     describe '異常系' do
+      invalid_attribute = {
+        category_id: ['0' * 33, 'g' * 32],
+      }
+
       it_behaves_like '必須パラメーターがない場合のテスト', %i[name]
+      it_behaves_like '不正な値を指定した場合のテスト', invalid_attribute
     end
   end
 end