+++ spec/models/payment_spec.rb
@@ -6,6 +6,7 @@ describe Payment, type: :model do
   describe '#validates' do
     describe '正常系' do
       valid_attribute = {
+        payment_id: ['0' * 32],
         payment_type: %w[income expense],
         date: %w[1000-01-01 1000/01/01 01-01-1000 01/01/1000 10000101],
         content: 'モジュールテスト用データ',
@@ -17,6 +18,7 @@ describe Payment, type: :model do
 
     describe '異常系' do
       invalid_attribute = {
+        payment_id: ['0' * 33, 'g' * 32],
         payment_type: 'invalid',
         date: %w[invalid 1000-13-01 1000-01-00 1000-13-00],
         price: [-1],