+++ CHANGELOG.md
@@ -1,3 +1,16 @@
+# 4.5.0 (2020/03/07)
+- [ADD] view to manage tags and categories
+- [UPDATE] whole view design
+- [UPDATE] gems
+
+# 4.4.1 (2020/02/09)
+- [UPDATE] resources to add resource id
+- [UPDATE] gems
+
+# 4.4.0 (2019/01/13)
+- [UPDATE] payment resource to set tags
+- [UPDATE] gems
+
 # 4.3.2 (2019/12/02)
 - [UPDATE] ruby version to 2.6.3
 - [UPDATE] gems